requirements.txt
# AMC Clinical - Main Requirements
# Core database and PDF processing

# PDF Processing
pypdf==6.0.0

# Text Processing
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.9

# AI and LLM Integration
langchain-openai==0.3.32
langchain-pinecone==0.2.11
openai>=1.0.0
pinecone==7.3.0

# Environment and Configuration
python-dotenv==1.0.0

# Django REST Framework
djangorestframework==3.15.2

# Additional utilities
uuid==1.30
